+++ primefaces/src/main/java/org/primefaces/component/datatable/export/DataTablePDFExporter.java
@@ -23,15 +23,14 @@
  */
 package org.primefaces.component.datatable.export;
 
-import java.awt.*;
-import java.io.IOException;
-import java.util.Objects;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.logging.Logger;
-import javax.faces.context.FacesContext;
-
+import com.lowagie.text.Document;
+import com.lowagie.text.DocumentException;
+import com.lowagie.text.Element;
 import com.lowagie.text.Font;
-import com.lowagie.text.*;
+import com.lowagie.text.FontFactory;
+import com.lowagie.text.PageSize;
+import com.lowagie.text.Paragraph;
+import com.lowagie.text.Phrase;
 import com.lowagie.text.pdf.BaseFont;
 import com.lowagie.text.pdf.PdfPCell;
 import com.lowagie.text.pdf.PdfPTable;
@@ -46,6 +45,12 @@ import org.primefaces.component.export.PDFOrientationType;
 import org.primefaces.util.Constants;
 import org.primefaces.util.IOUtils;
 import org.primefaces.util.LangUtils;
+
+import javax.faces.context.FacesContext;
+import java.awt.*;
+import java.io.IOException;
+import java.util.Objects;
+import java.util.logging.Logger;
 
 public class DataTablePDFExporter extends DataTableExporter<Document, PDFOptions> {
 
@@ -123,14 +128,6 @@ public class DataTablePDFExporter extends DataTableExporter<Document, PDFOptions
     @Override
     protected void exportColumnFacetValue(FacesContext context, DataTable table, ColumnValue columnValue, int index) {
         addFacetValue(1, 1, columnValue);
-    }
-
-    @Override
-    protected void exportColumnGroupFacetValueLegacy(FacesContext context, DataTable table, UIColumn column,
-                                               AtomicInteger colIndex, ColumnValue columnValue) {
-        int rowSpan = column.getExportRowspan() != 0 ? column.getExportRowspan() : column.getRowspan();
-        int colSpan = column.getExportColspan() != 0 ? column.getExportColspan() : column.getColspan();
-        addFacetValue(rowSpan, colSpan, columnValue);
     }
 
     @Override
