+++ primefaces/src/main/java/org/primefaces/component/subtable/SubTable.java
@@ -24,15 +24,12 @@
 package org.primefaces.component.subtable;
 
 import java.util.List;
-import java.util.Map;
 
 import javax.faces.event.PhaseId;
 
-import org.primefaces.component.api.ColumnAware;
 import org.primefaces.component.api.UIColumn;
-import org.primefaces.model.ColumnMeta;
 
-public class SubTable extends SubTableBase implements ColumnAware {
+public class SubTable extends SubTableBase {
 
     public static final String COMPONENT_TYPE = "org.primefaces.component.SubTable";
 
@@ -58,10 +55,5 @@ public class SubTable extends SubTableBase implements ColumnAware {
     @Override
     public void setColumns(List<UIColumn> columns) {
         this.columns = columns;
-    }
-
-    @Override
-    public void setColumnMeta(Map<String, ColumnMeta> columnMeta) {
-        // NOOP
     }
 }
